Added an image server scheme property to application.properties
Added image server scheme property to application.properties to make
local debugging easier and also fixed naming of HTTP HEADERS that are
set by the Spring application to NOT use underscore characters as those
are now silently ignored with Apache 2.4 - I actually added two new
headers, one with dashes (which are translated to underscores) by some
layer of the stack and another header with no delimiters which will be
the future migration header to use.  I needed backwards compatibility
with the existing system which is why I didn't just change it and be
done.