To support GuideKick mobile app - add media items and Linked Art support
Added the beginnings of support for the LinkedArt data modeling
standards.  There will be a lot more work to do on that front as the
RESTful API for LinkedArt hasn't even been discussed yet and many of the
data modeling standards are in flux.  I expect progress to be fairly
rapid though once that effort starts to take off again.

Added support for finding art objects by their on view within the
institution status (onView=true|false)as well as whether they have
overview text (hasOverviewText=true|false).

Added support for audio and video assets that refer to objects.

These enhancements were developed relatively quickly and the class
structures for the LinkedArt model need some refactoring down the line.
I also need to add a few more tests to support the new functionality
that was added but the existing integration tests are all passing.